fix: fix ImagUtil class bugs.

+++ app/Libraries/ImageUtil.php
@@ -2,29 +2,34 @@
 
 namespace App\Libraries;
 
+use Exception;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Storage;
-use Psy\Util\Str;
+use Illuminate\Support\Str;
+use League\Flysystem\UnableToWriteFile;
+use Symfony\Component\HttpFoundation\File\Exception\UploadException;
 
 class ImageUtil
 {
-    public static function save(UploadedFile $file, string $directtory): string
+    public static function save(UploadedFile $file, string $directory): string|false
     {
-        $fileName = basename($file->getClientOriginalName(), '.' . $file->getClientOriginalExtension());
-        $imageName = Str::slug($fileName . '-' . time()) . '.' . $file->getClientOriginalExtension();
-        $path = $file->storeAs($directtory, $imageName, 'public');
-        return Storage::url($path);
-    }
+        $path = $file->store($directory, 'public'); 
+        
+        return $path ? Storage::url($path) : false;
+    }   
 
-    public static function remove(?string $file): void
+    public static function remove(string $file): bool
     {
-        if (!$file) {
-            return;
+        if (!$file || !str_starts_with($file, '/storage/')) {
+            return false;
         }
 
         $path = str_replace('/storage/', '', $file);
-        if (Storage::disk('public')->exists($path)) {
-            Storage::disk('public')->delete($path);
+        
+        if (!Storage::disk('public')->exists($path)) {
+            return false;
         }
+        
+        return Storage::disk('public')->delete($path);
     }
 }
